New job log table view
