+++ src/views/jobManager/jobLog.ts
@@ -1,5 +1,6 @@
 import { ViewColumn, window } from "vscode";
 import { JobInfo } from "../../connection/manager";
+import { head } from "../html";
 
 interface JobLogMessage {
   MESSAGE_ID: string;
@@ -23,86 +24,44 @@ export async function displayJobLog(selected: JobInfo) {
 }
 
 function generatePage(rows: JobLogMessage[]) {
-  return /*html*/`
-<!DOCTYPE html>
-<html>
-  <head>
-    <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <style>
-      @media only screen and (min-width: 750px) {
-        #jobLogRows {
-          padding-left: 15%;
-          padding-right: 15%;
-        }
-      }
-      #jobLogRows {
-        margin: 2em 2em 2em 2em;
-      }
+  return /*html*/ `
+    <!DOCTYPE html>
+    <html lang="en">
+      <head>
+        ${head}
+        <script>
+          window.addEventListener("load", main);
+          function main() {}
+        </script>
+      </head>
+      <body>
+        <vscode-data-grid grid-template-columns="150px 100px auto">
+          ${rows.map(row => {
+            let i = 1;
 
+            // <vscode-data-grid-cell grid-column="${i++}">
+            //   ${row.MESSAGE_TIMESTAMP}
+            // </vscode-data-grid-cell>
+            // <vscode-data-grid-cell grid-column="${i++}">
+            //   <code>${row.FROM_LIBRARY}/${row.FROM_PROGRAM}</code>
+            // </vscode-data-grid-cell>
 
-      .toTheRight {
-        float: right;
-      }
-
-      .collapsible {
-        background-color: var(--vscode-button-secondaryBackground);
-        color: var(--vscode-foreground);
-        cursor: pointer;
-        padding: 1em;
-        width: 100%;
-        border: none;
-        outline: none;
-        font-size: 15px;
-        margin-top: 2px;
-        margin-bottom: 2px;
-        text-align: left;
-      }
-      .active,
-      .collapsible:hover {
-        background-color: var(--vscode-button-secondaryHoverBackground);
-      }
-
-      .content {
-        padding: 0 18px;
-        display: none;
-        overflow: hidden;
-        background-color: var(--vscode-button-secondaryHoverBackground);
-        margin-left: 3em;
-        margin-top: 2px;
-        margin-bottom: 2px;
-      }
-    </style>
-  </head>
-  <body>
-    <div id="jobLogRows">
-      ${rows.map(row => 
-        `<button type="button" class="collapsible">[<code>${row.MESSAGE_ID}</code>] ${escapeHTML(row.MESSAGE_TYPE)} <span class="toTheRight">${escapeHTML(row.MESSAGE_TIMESTAMP)}</span></button>
-        <div class="content">
-          <p>${row.FROM_LIBRARY}/${row.FROM_PROGRAM}</p>
-          <p>${escapeHTML(row.MESSAGE_TEXT)}</p>
-        </div>`
-        ).join(``)}
-    </div>
-
-    <script>
-      var coll = document.getElementsByClassName("collapsible");
-      var i;
-
-      for (i = 0; i < coll.length; i++) {
-        coll[i].addEventListener("click", function () {
-          this.classList.toggle("active");
-          var content = this.nextElementSibling;
-          if (content.style.display === "block") {
-            content.style.display = "none";
-          } else {
-            content.style.display = "block";
-          }
-        });
-      }
-    </script>
-  </body>
-</html>
-  `
+            return `<vscode-data-grid-row>
+              <vscode-data-grid-cell grid-column="${i++}">
+                ${row.MESSAGE_TYPE}
+              </vscode-data-grid-cell>
+              <vscode-data-grid-cell grid-column="${i++}">
+                ${row.MESSAGE_ID}
+              </vscode-data-grid-cell>
+              <vscode-data-grid-cell grid-column="${i++}">
+                ${escapeHTML(row.MESSAGE_TEXT)}
+              </vscode-data-grid-cell>
+            </vscode-data-grid-row>`
+          }).join(``)}
+        </vscode-data-grid>
+      </body>
+    </html>
+  `;
 }
 
 const escapeHTML = str => str.replace(/[&<>'"]/g,
